Generates list of options for a SAGA-GIS tool

Parses the html table for a SAGA-GIS tool into a list of identifiers, options, defaults and constraints

create_tool(tool_information, tool_options, description, html_file)

Arguments

tool_information

list

tool_options

list

description

the description text for the tool that has been scraped from the help documentation

html_file

the name of the html file for the tool's documentation. Stored to help linking with online documentation.

Value

A `saga_tool` object containing: + `tool_name` A syntactically-correct name for the tool. + `description` The tool's description. + `author` The tool's author. + `tool_cmd` The command to use for saga_cmd to execute tool. + `tool_id` The tool's ID. + `parameters` A named list of the tool's parameter objects. + `html_file` The html document name.