Apply a saga_cmd tool over over a list of raster datasets and mosaick the results

A `purrr::map` style function to iterate over a list of raster datasets and apply a saga_cmd tool to each raster. The output grids are automatically mosaicked into a new raster dataset. The `map_raster` function is can help with conveniently applying terrain analysis functions to lists of rasters. One potential application is using tiling to deal with raster datasets that are too large to fit into memory, and when the `saga_gis(grid_caching = TRUE)` option is too slow. The tiles can potentially generated using the `Rsagacmd::tile_geoprocessor` tool, or other functions that split a dataset. pck

map_raster(.x, .f, filename = NULL, options = NULL, nodata = NULL)

Arguments

.x
A list.

.f
A saga_cmd function that was generated by `Rsagacmd::saga_gis()` to apply to each raster dataset in `.x`. The function can be used as it, such as `saga$ta_morphometry$terrain_ruggedness_index` or can be wrapped inside another as a partial function to set arguments to the tool or select specific tool outputs.

filename
A character to optionally specify the path to save the final mosaicked result. If `filename = NULL` then `tempdir()` is used.

options
A character vector of GDAL creation options used when saving the dataset, such as `options = c("COMPRESS=DEFLATE", "BIGTIFF=YES")`.

nodata
Any number to use as the nodata value for the output raster. If `nodata = NULL` then the nodata value will be based on the input raster's nodata values.

Value

Either a `raster::RasterLayer` or `terra::SpatRaster` object.

Examples

if (FALSE) {
library(Rsagacmd)
library(terra)

saga <- saga_gis(backend = "terra")

# generate a dem and split into tiles
dem <- saga$grid_calculus$random_terrain()
dem_tiles <- tile_geoprocessor(saga, dem, nx = 25, ny = 25)

# apply a saga tool function to each tile
tile_output <- map_raster(
    dem_tiles,
    saga$ta_morphometry$terrain_ruggedness_index
)

# alternatively use a partial function to set arguments to the tool
tri_func <- function(dem) {
    saga$ta_morphometry$terrain_ruggedness_index(
        dem = dem,
        radius = 7
     )
}

tile_output <- map_raster(dem_tiles, tri_func)

# if the saga tool produces multiple outputs then each output will be
# mosaicked automatically and returned as a list of raster datasets
tile_output <- map_raster(
    dem_tiles,
    saga$ta_hydrology$saga_wetness_index
)
}