Generic function to display help and usage information for any SAGA-GIS tool

Displays a tibble containing the name of the tool's parameters, the argument name used by Rsagacmd, the identifier used by the SAGA-GIS command line, and additional descriptions, default and options/constraints.

# S3 method for saga_tool
print(x, ...)

Arguments

x

A `saga_tool` object.

...

Additional arguments to pass to print. Currently not used.

Examples

# \dontrun{
# Initialize a saga object
saga <- saga_gis()

# Display usage information on a tool
print(saga$ta_morphometry$slope_aspect_curvature)
#> Help for library = ta_morphometry; tool = slope_aspect_curvature:
#> Author: nO.Conrad (c) 2001 
#> Description: Calculates the local morphometric terrain parameters slope, aspect and if supported by the chosen method also the curvature. Besides tangential curvature also its horizontal and vertical components (i.e. plan and profile curvature) can be calculated. 
#> 
#> # A tibble: 16 × 7
#>    parameter        type  argument identifier description default available_opts
#>    <chr>            <chr> <chr>    <chr>      <chr>         <dbl> <chr>         
#>  1 Elevation        Grid… elevati… ELEVATION  NA               NA NA            
#>  2 Slope            Grid… slope    SLOPE      NA               NA NA            
#>  3 Aspect           Grid… aspect   ASPECT     NA               NA NA            
#>  4 General Curvatu… Grid… c_gene   C_GENE     NA               NA NA            
#>  5 Profile Curvatu… Grid… c_prof   C_PROF     NA               NA NA            
#>  6 Plan Curvature   Grid… c_plan   C_PLAN     NA               NA NA            
#>  7 Tangential Curv… Grid… c_tang   C_TANG     NA               NA NA            
#>  8 Longitudinal Cu… Grid… c_long   C_LONG     Zevenberge…      NA NA            
#>  9 Cross-Sectional… Grid… c_cros   C_CROS     Zevenberge…      NA NA            
#> 10 Minimal Curvatu… Grid… c_mini   C_MINI     NA               NA NA            
#> 11 Maximal Curvatu… Grid… c_maxi   C_MAXI     NA               NA NA            
#> 12 Total Curvature  Grid… c_tota   C_TOTA     NA               NA NA            
#> 13 Flow Line Curva… Grid… c_roto   C_ROTO     NA               NA NA            
#> 14 Method           Choi… method   METHOD     NA                6 [0] maximum s…
#> 15 Unit             Choi… unit_sl… UNIT_SLOPE NA                0 [0] radians; …
#> 16 Unit             Choi… unit_as… UNIT_ASPE… NA                0 [0] radians; …

# Or simply:
saga$ta_morphometry$slope_aspect_curvature
#> Help for library = ta_morphometry; tool = slope_aspect_curvature:
#> Author: nO.Conrad (c) 2001 
#> Description: Calculates the local morphometric terrain parameters slope, aspect and if supported by the chosen method also the curvature. Besides tangential curvature also its horizontal and vertical components (i.e. plan and profile curvature) can be calculated. 
#> 
#> # A tibble: 16 × 7
#>    parameter        type  argument identifier description default available_opts
#>    <chr>            <chr> <chr>    <chr>      <chr>         <dbl> <chr>         
#>  1 Elevation        Grid… elevati… ELEVATION  NA               NA NA            
#>  2 Slope            Grid… slope    SLOPE      NA               NA NA            
#>  3 Aspect           Grid… aspect   ASPECT     NA               NA NA            
#>  4 General Curvatu… Grid… c_gene   C_GENE     NA               NA NA            
#>  5 Profile Curvatu… Grid… c_prof   C_PROF     NA               NA NA            
#>  6 Plan Curvature   Grid… c_plan   C_PLAN     NA               NA NA            
#>  7 Tangential Curv… Grid… c_tang   C_TANG     NA               NA NA            
#>  8 Longitudinal Cu… Grid… c_long   C_LONG     Zevenberge…      NA NA            
#>  9 Cross-Sectional… Grid… c_cros   C_CROS     Zevenberge…      NA NA            
#> 10 Minimal Curvatu… Grid… c_mini   C_MINI     NA               NA NA            
#> 11 Maximal Curvatu… Grid… c_maxi   C_MAXI     NA               NA NA            
#> 12 Total Curvature  Grid… c_tota   C_TOTA     NA               NA NA            
#> 13 Flow Line Curva… Grid… c_roto   C_ROTO     NA               NA NA            
#> 14 Method           Choi… method   METHOD     NA                6 [0] maximum s…
#> 15 Unit             Choi… unit_sl… UNIT_SLOPE NA                0 [0] radians; …
#> 16 Unit             Choi… unit_as… UNIT_ASPE… NA                0 [0] radians; …
# }