Search for a SAGA-GIS tool

Search for a SAGA-GIS tool

search_tools(x, pattern)

Arguments

x

saga object

pattern

character, pattern of text to search for within the tool name

Value

a tibble containing the libraries, names and parameters of the tools that match the pattern of the search text and their host library

Examples

# \dontrun{
# initialize Rsagacmd
saga <- saga_gis()

# search for a tool
search_tools(x = saga, pattern = "terrain")
#> # A tibble: 9 × 6
#>   library            tool                 author saga_cmd parameters description
#>   <chr>              <chr>                <chr>  <chr>    <list>     <chr>      
#> 1 grid_calculus      random_terrain       V.Ola… Random … <chr [13]> (c) 2004 b…
#> 2 grid_visualisation terrain_map_view     Copyr… Terrain… <chr [9]>  This tool …
#> 3 io_shapes          import_wasp_terrain… O.Con… Import … <chr [3]>  Import WAs…
#> 4 ta_compound        basic_terrain_analy… O.Con… Basic T… <chr [18]> A selectio…
#> 5 ta_morphometry     terrain_ruggedness_… O.Con… Terrain… <chr [7]>  Terrain Ru…
#> 6 ta_morphometry     terrain_surface_tex… O.Con… Terrain… <chr [8]>  Terrain su…
#> 7 ta_morphometry     terrain_surface_con… O.Con… Terrain… <chr [10]> Terrain su…
#> 8 ta_morphometry     terrain_surface_cla… O.Con… Terrain… <chr [14]> Terrain su…
#> 9 terrain_analysis   terrain_clustering   O.Con… Terrain… <chr [4]>  created fr…
# }