Summarize the libraries that are available within a saga object and return these as a tibble.

Summarize the libraries that are available within a saga object and return these as a tibble.

# S3 method for saga
tidy(x, ...)

Arguments

x

a `saga` object

...

additional arguments. Currently unused.

Value

a tibble that describes libraries, their descriptions and number of tools that are available in SAGA-GIS.

Examples

# \dontrun{
# Initialize a saga object
saga <- saga_gis()

# tidy the saga object's parameters into a tibble
tidy(saga)
#> # A tibble: 67 × 3
#>    libraries       description                                           n_tools
#>    <chr>           <chr>                                                   <int>
#>  1 Grid Tools      No description                                              1
#>  2 Polygon Tools   No description                                              2
#>  3 climate_tools   Tools for the processing and analysis of climate and…      29
#>  4 contrib_perego  Contributions from alessandro perego. Go to for furt…       8
#>  5 garden_fractals Fractals                                                    4
#>  6 gc_tools        Tools for grid collections.                                 5
#>  7 grid_analysis   Some grid analysis tools.                                  23
#>  8 grid_calculus   NA                                                         21
#>  9 grid_filter     Filters for grids.                                         22
#> 10 grid_gridding   Tools for the gridding of points and other vector da…      11
#> # … with 57 more rows
# }